refactor: Move tools to individual install scripts
After a failed Neovim install stopped the entire install
script, it became obvious that I needed to have a way to
march on after failures. I think it's nice for each item
I want to install to have its own script so I can go back
and just run that script instead of the the whole kit and
caboodle, even if it is idempotent.